+++ packages/cli/src/commands/respect/har-logs/helpers/get-input-url.ts
@@ -1,4 +1,0 @@
-export function getInputUrl(input: any): URL {
-  const url = typeof input === 'string' ? input : input.url;
-  return new URL(url);
-}