+++ docs/@v2/changelog.md
@@ -7,6 +7,44 @@ toc:
 
 <!-- do-not-remove -->
 
+## 2.15.1 (2026-02-02)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.15.1.
+
+## 2.15.0 (2026-01-27)
+
+### Patch Changes
+
+- Removed unused `chokidar` dependency.
+- Updated @redocly/openapi-core to v2.15.0.
+
+## 2.14.9 (2026-01-23)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.14.9.
+
+## 2.14.8 (2026-01-23)
+
+### Patch Changes
+
+- Fixed an issue where multi-line lint error messages could break Markdown formatting.
+- Updated @redocly/openapi-core to v2.14.8.
+
+## 2.14.7 (2026-01-22)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.14.7.
+
+## 2.14.6 (2026-01-21)
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.14.6.
+
 ## 2.14.5 (2026-01-12)
 
 ### Patch Changes